Got the controller working, but not the handlers clients
