+++ pkg/apis/management.cattle.io/v3/global_types.go
@@ -21,6 +21,7 @@ type Setting struct {
 
 // +genclient
 // +genclient:nonNamespaced
+// +kubebuilder:resource:scope=Cluster
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
 
 type Feature struct {
